requirements2.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# PDF processing
PyPDF2>=3.0.1

# LangChain ecosystem
langchain>=0.1.0
langchain-huggingface>=0.0.8
faiss-cpu>=1.7.4

# NLP & Embeddings
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0

# Optional (for OpenAI embeddings if needed)
openai>=1.6.0
